Move analytics to own src/ level module
- refactoring commit
