+++ src/options/imports/actions.js
@@ -1,6 +1,6 @@
 import { createAction } from 'redux-act'
 
-import analytics from 'src/util/analytics'
+import analytics from 'src/analytics'
 import db from 'src/pouchdb'
 import { CMDS, IMPORT_CONN_NAME, OLD_EXT_KEYS } from './constants'
 import * as selectors from './selectors'
