+++ src/overview/actions.js
@@ -1,6 +1,6 @@
 import { createAction } from 'redux-act'
 
-import analytics from 'src/util/analytics'
+import analytics from 'src/analytics'
 import { remoteFunction } from 'src/util/webextensionRPC'
 import * as constants from './constants'
 import * as selectors from './selectors'
