+++ src/popup/components/LinkButton.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import analytics from 'src/util/analytics'
+import analytics from 'src/analytics'
 import Button from './Button'
 
 import styles from './Button.css'
